Add script tree logic to ScriptTree and controller buttons
